requirements.txt
#### DO NOT CHANGE THIS WITHOUT ALSO CHANGING THE `grotto-deps` REPO

Django==3.1.4
django-appconf==1.0.4
django-colorfield==0.4.1
django-compressor==2.4
django-libsass==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
hashids==1.3.1
libsass==0.20.1
markovify==0.8.3
numpy==1.20.1
spacy==3.0.1
spacy-legacy==3.0.1
sqlparse==0.4.1
Unidecode==1.2.0
urllib3==1.26.3
django-enumfield
webcolors==1.11.1

# deployment needs
gunicorn==20.0.4
whitenoise==5.2.0

# REST API
djangorestframework==3.12.4
Markdown==3.3.4
django-filter==2.4.0
drf-yasg==1.20.0
